implement methods from SAF to Service class
