+++ src/main/java/com/cbfacademy/apiassessment/SelfAssessmentServiceInterface.java
@@ -1,0 +1,11 @@
+
+package com.cbfacademy.apiassessment;
+
+// create interface for self assessment service for api class . 
+
+// can we implement self assessment functions in here? 
+
+public interface SelfAssessmentServiceInterface {
+
+    
+}
