--- src/main/java/com/cbfacademy/apiassessment/StoringSelfAssessmentData.java
@@ -1,17 +1,0 @@
-package com.cbfacademy.apiassessment;
-
-public class StoringSelfAssessmentData {
-
-//     public void saveSelfAssessmentDataToJSON(SelfAssessmentUserDetails dataFromSelfAssessmentInput) {
-
-//         Gson gson = new GsonBuilder().setPrettyPrinting().create(); 
-
-//         String gson.toJson(dataFromSelfAssessmentInput);
-
-//         try (FileWriter writer )
-
-
-
-//     }
-    
- }
